+++ packages/iov-ledger-bns/types/index.d.ts
@@ -1,2 +1,2 @@
-export { LedgerSimpleAddressKeyringEntry } from "./ledgersimpleaddresskeyringentry";
+export { LedgerSimpleAddressWallet } from "./ledgersimpleaddresskeyringentry";
 export { LedgerState } from "./statetracker";